webull quote converter for stock_indicators
new function to convert a webull quote into a dataframe which the stock_indicators package in python will accept.

I'm not sure how many people will actually use it, it they use it at all so I have imports inside of this function, instead of outside where people who don't even use it will need to install it. There's probally a better way to implement this but I'm not sure how to do that. 

To use this, it requires pandas and stock_indicators package to be installed. thes second package is linked here:

https://python.stockindicators.dev/
https://github.com/DaveSkender/Stock.Indicators.Python